6-RAG/CASE-ChatPDF-Faiss-250726/requirements.txt
langchain==0.1.0
langchain-community==0.0.11
faiss-cpu==1.7.4
dashscope==1.14.2
python-dotenv==1.0.1
pyprojroot==0.3.0
pandas==2.2.2
flask==2.3.2
flask-cors==4.0.0